+++ packages/mars-build/src/index.js
@@ -6,11 +6,12 @@
 const {
     build,
     clean,
-    watch,
-    merge
+    watch
 } = require('./scripts/run');
+
+const getConfig = require('./scripts/getConfig');
 
 exports.build = build;
 exports.clean = clean;
 exports.watch = watch;
-exports.merge = merge;
+exports.getConfig = getConfig;