+++ packages/mars-cli/lib/build.js
@@ -11,25 +11,30 @@ const execa = require('execa');
 const {getConfig} = require('./scripts/getConfig');
 
 async function build(cmd) {
-    const {config, target, buildPath} = getConfig(cmd);
+    const {target, buildPath} = getConfig(cmd);
 
     const {
-        build
-    } = require(buildPath + '/src/scripts/run');
+        build,
+        clean
+    } = require(buildPath);
 
-
-    process.env.NODE_ENV = 'production';
-
-    build(config, {
+    const options = {
         target,
         build: true
-    }).on('stop', () => {
-        if (target === 'h5') {
-            const child = execa('npm', ['run', 'build-dist-h5']);
-            child.stdout.pipe(process.stdout);
-            child.stderr.pipe(process.stderr);
-        }
+    };
+    process.env.NODE_ENV = 'production';
+    process.env.MARS_CLI_OPTIONS = JSON.stringify(options);
+
+    clean(options).once('stop', () => {
+        build(options).once('stop', () => {
+            if (target === 'h5') {
+                const child = execa('npm', ['run', 'build-dist-h5']);
+                child.stdout.pipe(process.stdout);
+                child.stderr.pipe(process.stderr);
+            }
+        });
     });
+
 }
 
 module.exports = (...args) =>