+++ docs/CHANGELOG.md
@@ -1,16 +1,42 @@
 # CHANGELOG
 
-### @marsjs/build@0.2.1 (20190404)
-- Initial Version
 
-### @marsjs/core@0.2.1 (20190404)
-- Initial Version
+## 2019-04-12
 
-### @marsjs/api@1.0.13 (20190404)
-- Initial Version
+- @marsjs/api@1.0.16
+- @marsjs/build@0.2.6
+- @marsjs/cli@0.1.17
+- @marsjs/cli-template@0.1.5
 
-### @marsjs/components@1.0.16 (20190404)
-- Initial Version
+### Bug Fixes
 
-### @marsjs/cli@0.1.12 (20190404)
-- Initial Version
+* **cli-template:** add default MARS_CLI_OPTIONS for mars-cli-service ([9b06d0d](https://github.com/max-team/Mars/commit/9b06d0d))
+* **build:** disable px2units for h5 App Style ([38f7996](https://github.com/max-team/Mars/commit/38f7996))
+* **build:** page 中 config 可能不存在 ([fca5a1f](https://github.com/max-team/Mars/commit/fca5a1f))
+* **build:** preprocessors merge 错误，不能设置 options ([b54850c](https://github.com/max-team/Mars/commit/b54850c))
+
+
+### Features
+
+* **cli:** set process.env.NODE_ENV ([f22c7ca](https://github.com/max-team/Mars/commit/f22c7ca))
+* **cli:** 编译前先清空编译产物 ([4d79b1e](https://github.com/max-team/Mars/commit/4d79b1e), [cf7bf84](https://github.com/max-team/Mars/commit/cf7bf84))
+* **build:** js 编译增加 process.env.NODE_ENV 替换 ([0e8fea1](https://github.com/max-team/Mars/commit/0e8fea1))
+* **build:** mars-cli info ([16cc438](https://github.com/max-team/Mars/commit/16cc438))
+* **build:** mars-cli update ([d6863d1](https://github.com/max-team/Mars/commit/d6863d1))
+* **build:** move config file parse to @marsjs/build ([c906d5e](https://github.com/max-team/Mars/commit/c906d5e))
+* **build/cli-template:** pass postcss config to vue-cli postcss-loader ([d66d4f7](https://github.com/max-team/Mars/commit/d66d4f7))
+
+* **build:** support config api/components packages ([842848c](https://github.com/max-team/Mars/commit/842848c))
+* **api:** update api ([a3017d2](https://github.com/max-team/Mars/commit/a3017d2), [5de627c](https://github.com/max-team/Mars/commit/5de627c), [d63104d](https://github.com/max-team/Mars/commit/d63104d))
+* **build:**  add page api onPageScroll and onTabItemTap ([#6](https://github.com/max-team/Mars/issues/6)) ([482e39c](https://github.com/max-team/Mars/commit/482e39c))
+* **build:** 支持使用原生小程序组件和页面 ([d62ecce](https://github.com/max-team/Mars/commit/d62ecce))
+
+## 2019-04-09
+
+- @marsjs/api@1.0.15
+- @marsjs/build@0.2.3
+- @marsjs/core@0.2.3
+- @marsjs/cli@0.1.14
+- @marsjs/cli-template@0.1.4
+
+